// src/app/components/model/produto.model.ts
export interface Produto {
    id?: string;
    nome?: String;
    ano?: number;
    modelo?: String;
    motor?: String;
    numeroDeMotor?: number;
    cor?: String;
    referencia?: String;
    marca?: String;
    lado?: String;
    dianteira?: String;
    traseira?: String;
    foto?: String;
    valor?: number;
    desconto?: number;
    quantidade?: bigint;
}